rename logger2
