+++ extensions/warp-blink-wrtc/src/lib.rs
@@ -10,7 +10,7 @@
 #![allow(dead_code)]
 
 mod host_media;
-mod rtp_logger2;
+mod rtp_logger;
 mod signaling;
 mod simple_webrtc;
 mod store;
@@ -119,7 +119,7 @@ impl Drop for BlinkImpl {
                 log::error!("error in webrtc_controller deinit: {e}");
             }
             host_media::reset().await;
-            rtp_logger2::de_init();
+            rtp_logger::de_init();
             log::debug!("deinit finished");
         });
     }
@@ -263,7 +263,7 @@ impl BlinkImpl {
             None => bail!("blink not initialized"),
         };
         // todo: don't faile because of this
-        rtp_logger2::init(call.call_id(), "/tmp/rtp-logs".into())?;
+        rtp_logger::init(call.call_id(), "/tmp/rtp-logs".into())?;
         self.active_call.write().await.replace(call.clone().into());
         let audio_source_codec = self.audio_source_codec.read().await;
         // ensure there is an audio source track
